+++ backend/models/__init__.py
@@ -9,7 +9,6 @@ from .settings import (BrainRateLimiting, BrainSettings, ContactsSettings,
                        LLMSettings, ResendSettings, get_embeddings,
                        get_documents_vector_store, get_embeddings,
                        get_supabase_client, get_supabase_db)
-from .user_identity import UserIdentity
 from .user_usage import UserUsage
 
 # TODO uncomment the below import when start using SQLalchemy
