+++ backend/modules/user/repository/create_user_identity.py
@@ -1,7 +1,8 @@
 from typing import Optional
 from uuid import UUID
 
-from models import UserIdentity, get_supabase_client
+from models import get_supabase_client
+from modules.user.entity.user_identity import UserIdentity
 
 
 def create_user_identity(id: UUID, openai_api_key: Optional[str]) -> UserIdentity:
@@ -19,5 +20,5 @@ def create_user_identity(id: UUID, openai_api_key: Optional[str]) -> UserIdentit
     )
     user_identity = response.data[0]
     return UserIdentity(
-        id=user_identity.get('user_id'), openai_api_key=user_identity.get('openai_api_key')  # type: ignore
+        id=user_identity.get("user_id"), openai_api_key=user_identity.get("openai_api_key")  # type: ignore
     )
