+++ backend/modules/user/repository/get_user_identity.py
@@ -1,8 +1,9 @@
 from multiprocessing import get_logger
 from uuid import UUID
 
-from models import get_supabase_client, UserIdentity
-from repository.user_identity.create_user_identity import create_user_identity
+from models import get_supabase_client
+from modules.user.entity.user_identity import UserIdentity
+from modules.user.repository import create_user_identity
 
 logger = get_logger()
