+++ backend/modules/user/repository/update_user_properties.py
@@ -2,10 +2,9 @@ from typing import Optional
 from uuid import UUID
 
 from models.settings import get_supabase_client
-from models.user_identity import UserIdentity
+from modules.user.entity.user_identity import UserIdentity
+from modules.user.repository import create_user_identity
 from pydantic import BaseModel
-
-from repository.user_identity.create_user_identity import create_user_identity
 
 
 class UserUpdatableProperties(BaseModel):
