+++ backend/modules/user/service/__init__.py
@@ -1,0 +1,1 @@
+from .get_user_id_by_email import get_user_id_by_email
