+++ backend/modules/user/service/get_user_id_by_email.py
@@ -1,0 +1,7 @@
+from uuid import UUID
+
+from modules.user.repository import get_user_id_by_user_email
+
+
+def get_user_id_by_email(email: str) -> UUID | None:
+    return get_user_id_by_user_email(email)
