+++ backend/packages/utils/__init__.py
@@ -1,1 +1,2 @@
 from .handle_request_validation_error import handle_request_validation_error
+from .parse_message_time import parse_message_time
