+++ backend/repository/brain/get_default_user_brain_or_create_new.py
@@ -1,11 +1,8 @@
-from models import BrainEntity, UserIdentity
+from models import BrainEntity
 from models.databases.supabase.brains import CreateBrainProperties
-from routes.authorizations.types import RoleEnum
-
+from modules.user.entity.user_identity import UserIdentity
 from repository.brain import create_brain, create_brain_user, get_user_default_brain
-from repository.brain.create_brain import create_brain
-from repository.brain.create_brain_user import create_brain_user
-from repository.brain.get_default_user_brain import get_user_default_brain
+from routes.authorizations.types import RoleEnum
 
 
 def get_default_user_brain_or_create_new(user: UserIdentity) -> BrainEntity:
