--- backend/repository/user/__init__.py
@@ -1,2 +1,0 @@
-from .get_user_email_by_user_id import get_user_email_by_user_id
-from .get_user_id_by_user_email import get_user_id_by_user_email
