+++ backend/routes/api_key_routes.py
@@ -6,7 +6,8 @@ from asyncpg.exceptions import UniqueViolationError
 from fastapi import APIRouter, Depends
 from logger import get_logger
 from middlewares.auth import AuthBearer, get_current_user
-from models import UserIdentity, get_supabase_db
+from models import get_supabase_db
+from modules.user.entity.user_identity import UserIdentity
 from pydantic import BaseModel
 
 logger = get_logger(__name__)
