+++ backend/routes/chat/utils.py
@@ -2,8 +2,9 @@ import time
 from uuid import UUID
 
 from fastapi import HTTPException
-from models import UserIdentity, UserUsage
+from models import UserUsage
 from models.databases.supabase.supabase import SupabaseDB
+from modules.user.entity.user_identity import UserIdentity
 
 
 class NullableUUID(UUID):
