+++ backend/routes/chat_routes.py
@@ -7,16 +7,9 @@ from fastapi.responses import StreamingResponse
 from llm.qa_base import QABaseBrainPicking
 from llm.qa_headless import HeadlessQA
 from middlewares.auth import AuthBearer, get_current_user
-from models import (
-    Brain,
-    BrainEntity,
-    Chat,
-    ChatQuestion,
-    UserIdentity,
-    UserUsage,
-    get_supabase_db,
-)
+from models import Brain, BrainEntity, Chat, ChatQuestion, UserUsage, get_supabase_db
 from models.databases.supabase.chats import QuestionAndAnswer
+from modules.user.entity.user_identity import UserIdentity
 from repository.chat import (
     ChatUpdatableProperties,
     CreateChatProperties,
