+++ backend/routes/onboarding_routes.py
@@ -1,13 +1,13 @@
 from fastapi import APIRouter, Depends
-from middlewares.auth import (
+from middlewares.auth import (  # Assuming you have a get_current_user function
+    AuthBearer,
     get_current_user,
-)  # Assuming you have a get_current_user function
-from middlewares.auth import AuthBearer
+)
 from models.databases.supabase.onboarding import (
     OnboardingStates,
     OnboardingUpdatableProperties,
 )
-from models.user_identity import UserIdentity
+from modules.user.entity.user_identity import UserIdentity
 from repository.onboarding.get_user_onboarding import get_user_onboarding
 from repository.onboarding.update_user_onboarding import update_user_onboarding
