+++ scripts/tables.sql
@@ -220,15 +220,6 @@ CREATE TABLE IF NOT EXISTS api_brain_definition (
     secrets JSON
 );
 
-CREATE OR REPLACE FUNCTION public.get_user_email_by_user_id(user_id uuid)
-RETURNS TABLE (email text)
-SECURITY definer
-AS $$
-BEGIN
-  RETURN QUERY SELECT au.email::text FROM auth.users au WHERE au.id = user_id;
-END;
-$$ LANGUAGE plpgsql;
-
 
 CREATE OR REPLACE FUNCTION public.get_user_id_by_user_email(user_email text)
 RETURNS TABLE (user_id uuid)
@@ -445,7 +436,7 @@ $$;
 
 
 INSERT INTO migrations (name) 
-SELECT '20231107104700_setup_vault'
+SELECT '20231114162700_drop_get_user_email_by_user_id'
 WHERE NOT EXISTS (
-    SELECT 1 FROM migrations WHERE name = '20231107104700_setup_vault'
+    SELECT 1 FROM migrations WHERE name = '20231114162700_drop_get_user_email_by_user_id'
 );
